second part of tc homework
a function, which does nothing on page, but only in console(can be refactored into a rendering function if needed).
In index.ts there are type obj and a few objects which are of type obj. Those objects are formed into array and passed to a function along with key value of type string and newKeyValue of type string.
The function loops through all items of the passed array and checks if element has key equal to passed key, if yes than changes key value to the one provided, that's basically it. Check console to see the results of function(it has to log initial key value and updated key value and return copy of the array but with updated objects.

To start, please download files into a single folder and run index.html as live server.